refactor(sources): move sources related to sources folder
